+++ fili-core/src/main/java/com/yahoo/bard/webservice/async/jobs/stores/NoOpApiJobStore.java
@@ -6,6 +6,8 @@ import com.yahoo.bard.webservice.async.jobs.jobrows.JobRow;
 
 import rx.Observable;
 
+import java.util.Set;
+
 import javax.inject.Singleton;
 
 /**
@@ -28,6 +30,19 @@ public class NoOpApiJobStore implements ApiJobStore {
 
     @Override
     public Observable<JobRow> getAllRows() {
+        return Observable.empty();
+    }
+
+    /**
+     * This method ignores the filters and returns an empty Observable.
+     *
+     * @param jobRowFilter  A Set of JobRowFilters where each JobRowFilter contains the JobField to be
+     * filtered on, the filter operation and the values to be compared to.
+     *
+     * @return  An empty Observable.
+     */
+    @Override
+    public Observable<JobRow> getFilteredRows(Set<JobRowFilter> jobRowFilter) {
         return Observable.empty();
     }
 }